refactor: currency 패키지 이동
