+++ backend/src/main/java/hanglog/expense/domain/repository/CurrencyRepository.java
@@ -1,6 +1,6 @@
 package hanglog.expense.domain.repository;
 
-import hanglog.expense.domain.Currency;
+import hanglog.currency.domain.Currency;
 import java.time.LocalDate;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
