+++ backend/src/main/java/hanglog/expense/dto/response/TripExpenseResponse.java
@@ -2,8 +2,8 @@ package hanglog.expense.dto.response;
 
 
 import hanglog.city.dto.response.CityResponse;
+import hanglog.currency.domain.Currency;
 import hanglog.expense.domain.CategoryExpense;
-import hanglog.expense.domain.Currency;
 import hanglog.expense.domain.DayLogExpense;
 import hanglog.trip.domain.Trip;
 import hanglog.trip.domain.TripCity;
